completely revamp front-end with semantic ui and changes to db model
Front-end has been replaced with semantic ui to beautify and improve
overall ux. Most of the functionality and ui remains the same, with
minor changes to the database model, views and templating.

Most notably, an additional field "user type" has been added to the db
for scoping the views that users will be allowed to access later on.

In addition, the "register" view has been updated to reflect these changes
and the db migrations have been performed.

ToDo: important points to address before merging this branch with master
will be to reimplement the footer, which is in the process of being fixed
(did not include fix with this commit in case of breaking changes) and
update the current "base.html" to reflect this new fix so that it works
accross all views. In addition, an important change required is to fix the
"edit employees" view/route to pull in all available options for
populating the table in the view instead of simply the current options
that only applied to the particular user being modified. Finally,
redundant comments, logging etc should be removed. This is a dirty commit.